Relocate new classes implementing MonetaryOperator to function
