+++ src/main/java/org/javamoney/moneta/function/PrecisionContextRoundedOperator.java
@@ -1,4 +1,4 @@
-package org.javamoney.moneta;
+package org.javamoney.moneta.function;
 
 import java.math.BigDecimal;
 import java.math.MathContext;
@@ -7,6 +7,8 @@ import java.util.Objects;
 
 import javax.money.MonetaryAmount;
 import javax.money.MonetaryOperator;
+
+import org.javamoney.moneta.RoundedMoney;
 
 /**
  * <p>This implementation uses a {@link MathContext} to does the rounding operations. The implementation will use the <b>precision</b>, in other words, the total number of digits in a number</p>
